Update patch set 1
Patch Set 1:

In ONAP we hit a strange problem with building a binary distribution that might be due to an issue in Python 3.7.0, you can see the log here:

https://jenkins.onap.org/sandbox/job/dcaegen2-platform-configbinding-tox-sonar/6/console

This change upgrades pyenv installed Python 3.x distros to latest 2019 releases.  Please tell me if this is too bleeding edge or how you prefer to test & proceed, thanks.

Patch-set: 1
Label: Code-Review=0