Update patch set 1
Patch Set 1:

In terms of bleeding edge; python is not really semantically versioned, however for the patch versions it is. Meaning, they sometimes DO break stuff in o.X.o changes, but o.o.X changes are always bugfixes and patches. As a python developer, my recommendation is that switching any code project from o.X.o to o.>X.o requires investigation, testing, and should be completed with care, but upgrading to the latest patch should always work. My code is not on 3.7.4 yet but per the above principle I assume to upgrade and it'll work.

Patch-set: 1