Update patch set 1
Patch Set 1: Code-Review+1

I think we need to make sure that all of the global-jjb scripts we have that currently call pip natively are updated to `python -m pip` before we allow this to merge. So, I'm not giving it a +2 at this time, but it looks _correct_ to me.

Patch-set: 1
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=+1