Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> Shouldn't the var file name be ubuntu-20.04.json for consistency?

No because the current filter is checking for ansible_distro_version_major which is only the 20 in 20.04. If we changed the filter it would break the CentOS 8 file because that one is also filtering the same major version.

I'm simply reusing what's already in this project and not changing the code around the versioning.

Patch-set: 3