Fix: Install python 3.10.x on CentOS 7
Python 3.10.x is a pre-requiste for the ansible version 9.2.0
used fro the packer jobs. However the version of python 3.10.x
requires openssl11 libs and headers to be installed so that pyenv
install builds the required runtime dependencies.

ERROR: "ModuleNotFoundErrorr No module named '_ssl'"
"The Python ssl extension was not compiled. Missing the OpenSSL lib?"

Ref: https://github.com/pyenv/pyenv/wiki/Common-build-problems
Change-Id: Ib1c51882f057b7cf2c3a3c9307d524c9cf45b3a2
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>