Fix: Make "pyenv global" command work for all OSs
It is unclear why Ubuntu was set to use "system <version>" for the
global python version. While this command does not cause an error,
it effectively sets the global version to the system version, which
is not what it's supposed to accomplish.

In order to set the version to "<version>", we need to remove
"system" from the command. This makes the command for CentOS and
Ubuntu identical, so decision code has also been removed.

Change-Id: Ic232b7675e041f4c4aeb7cdcf1931dde0bebd113
Signed-off-by: Eric Ball <eball@linuxfoundation.org>