Fix tests and add Ubuntu 24.04 support
These needed to be in a single commit due to interdependent changes.

Since the old "Debian.yml" list was only supporting Ubuntu 18.04 at
this point, and the identical "Ubuntu-2{0,2}" files would have also
needed to be extended to 24.04, we simply swap newer versions to
Debian.yml, and the legacy 18.04 has its own file.

Additionally, the code block that was adding 3.10 and 3.11 python
versions to the list was not functioning as intended (the loop still
only installed 3.8 and 3.9, while 3.10 and 3.11 are installed in
separate tasks), and therefore was removed.

Change-Id: I0889499d11ba3b9b4b26be3ba205e2464675bdf0
Signed-off-by: Eric Ball <eball@linuxfoundation.org>